+++ app/controllers/sessions_controller.rb
@@ -129,12 +129,11 @@ class SessionsController < ApplicationController
       Rails.logger.error("Device authenticated successfully: #{session[:did]}")
       session.delete(:did)
 
-      redirect_to params[:url], allow_other_host: true, status: :found
-      head :no_content if performed?
-      # respond_to do |format|
-      #   format.turbo_stream { render turbo_stream: turbo_stream.replace("otp_input", partial: "sessions/success") }
-      #   format.html { redirect_to success_path }
-      # end
+      # redirect_to params[:url], allow_other_host: true, status: :found
+      respond_to do |format|
+        format.turbo_stream { render turbo_stream: turbo_stream.action(:redirect, params[:url]) }
+        format.html { redirect_to params[:url], allow_other_host: true, status: :found }
+      end
     else
       respond_to do |format|
         format.turbo_stream { render turbo_stream: turbo_stream.replace("otp_input", partial: "sessions/failed") }